+++ Tests/Fixtures/TestProject/StaticLibrary_Swift/StaticLibrary.swift
@@ -1,6 +1,6 @@
 
 public struct SLSwift {
     public func description() -> String {
-        return "Hello, World!"
+        "Hello, World!"
     }
 }